v11.0.0
* Breaking change: Response and Request were adjusted to enable duplication - getters and constructors provided.
* Breaking change: database commands "loadData" and "registerIndex" return types are
  their class (Db or DbEngine2), not their parent class (AbstractDb).
* New feature: possible to access the lock, "dbLock", which surrounds database change methods, and which
  can be used by developers to wrap multiple-statements, knowing that no other database calls can
  interfere while they run.  This is particularly important for "transactional" changes, where multiple
  changes have to be made.  This does not make the database ACID compliant, but it is a big step
  forwards in robustness.
* Correctness: Noticed a few classes were unintentionally left non-final.  Made final.
* Correctness: if the user tries registering multiple databases to the same path, an exception will be thrown.
* better logging, brig startup improved
  Noticed that the brig startup did not follow best practices, adjusted to fit.
* Removed superfluous locks around some methods in the brig, simplifying the code
* Allow the brig to return its database values more directly, avoiding unnecessary memory usage
* Fix: Realized that the dumpAttacker() method was too early if the client was using a keep-alive socket.
* More documentation
* In system constants, convert some ints to longs
* Stricter correctness handling on system constants

+++ src/main/java/com/renomad/minum/database/DatabaseAppender.java
@@ -81,7 +81,7 @@ final class DatabaseAppender {
     /**
      * This is the count of bytes that have been appended
      */
-    private long appendBytes;
+    private int appendBytes;
 
     DatabaseAppender(Path persistenceDirectory, Context context, IFileUtils fileUtils) throws IOException {
         this.fileUtils = fileUtils;
@@ -113,7 +113,7 @@ final class DatabaseAppender {
             appendCount = lines.size();
         } else {
             // reset the count to zero, we're starting a new file.
-            logger.logDebug(() -> "Creating a new database append file. Previous file: %,d lines, %.2f megabytes".formatted(appendCount, (appendBytes / 1_048_576.0)));
+            logger.logDebug(() -> "Creating a new database append file. Previous file: %,d lines, %.2f megabytes".formatted(appendCount, ((double) appendBytes / 1_048_576.0)));
             appendCount = 0;
             appendBytes = 0;
         }
@@ -166,7 +166,7 @@ final class DatabaseAppender {
                 // this code only runs when there is data to add, so no need to take a
                 // lot of waiting time.  But, if the data is coming fast and furious,
                 // at least a small wait will allow greater efficiency.
-                MyThread.sleep(50);
+                MyThread.sleep(50L);
             }
             flushLoopRunning = false;
         };
